+++ CHANGES.rst
@@ -1,6 +1,9 @@
 Next release
 ============
 
+With thanks to Feilong Ma for contributions.
+
+* [ENH] Option to provide a custom reference grid image (``--output-grid-reference``) for determining the field of view and resolution of output images (#480)
 * [ENH] Improved EPI skull stripping and tissue contrast enhancements (#519)
 * [ENH] Improve resource use estimates in FreeSurfer workflow (#506)
 * [ENH] Moved missing values in the DVARS* and FramewiseDisplacement columns of the _confounds.tsv from last row to the first row (#523)